split tournament details sections into components
